// packages/simmerbase/src/schemas/genera.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const GeneraBaseRowSchema = z.object({
	abbreviation: z.string(),
	description: z.string().nullable(),
	genus_name: z.string(),
	id: z.uuid(),
});

export type GeneraBaseRow = z.infer<typeof GeneraBaseRowSchema>;

export const GeneraInsertSchema = GeneraBaseRowSchema;

export type GeneraInsert = z.infer<typeof GeneraInsertSchema>;

export const GeneraUpdateSchema = GeneraInsertSchema.omit({ id: true }).partial();

export type GeneraUpdate = z.infer<typeof GeneraUpdateSchema>;