// packages/simmerbase/src/schemas/region_folders.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const RegionFoldersBaseRowSchema = z.object({
	created_at: z.coerce.date<Date>(),
	created_by: z.uuid().nullable(),
	group_id: z.uuid(),
	id: z.uuid(),
	region_folder_name: z.string(),
	updated_at: z.coerce.date<Date>(),
	updated_by: z.uuid().nullable(),
});

export type RegionFoldersBaseRow = z.infer<typeof RegionFoldersBaseRowSchema>;

export const RegionFoldersInsertSchema = z.object({
	group_id: z.uuid(),
	id: z.uuid(),
	region_folder_name: z.string(),
});

export type RegionFoldersInsert = z.infer<typeof RegionFoldersInsertSchema>;

export const RegionFoldersUpdateSchema = RegionFoldersInsertSchema.omit({ id: true }).partial();

export type RegionFoldersUpdate = z.infer<typeof RegionFoldersUpdateSchema>;