// packages/simmerbase/src/schemas/roles.ts
// Auto-generated by create-types-files.ts — 2026-02-13T00:41:02.350Z
// Do not edit manually.

import z from 'zod';

export const RolesBaseRowSchema = z.object({
	description: z.string().nullable(),
	id: z.number(),
	role_name: z.string(),
});

export type RolesBaseRow = z.infer<typeof RolesBaseRowSchema>;

export const RolesInsertSchema = RolesBaseRowSchema;

export type RolesInsert = z.infer<typeof RolesInsertSchema>;

export const RolesUpdateSchema = RolesInsertSchema.omit({ id: true }).partial();

export type RolesUpdate = z.infer<typeof RolesUpdateSchema>;